requirements.txt
# 亚马逊商品数据分析系统 - 依赖包

# 数据抓取
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# AI
anthropic>=0.18.0

# 数据库
# sqlite3 是Python内置模块，无需安装

# 模板和报告
jinja2>=3.1.0
plotly>=5.18.0

# 工具
python-dotenv>=1.0.0
openpyxl>=3.1.0

# 文本分析
scikit-learn>=1.3.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 日志
colorlog>=6.7.0

# 进程管理
psutil>=5.9.0

# WebSocket (用于卖家精灵)
websocket-client>=1.6.0
